Using exponential backoff in timeouts
